feat(results): fastatic returns filesizes in results (#92)
The results module returns an object with filesizes per directory per filetype.
The fastatic module returns this object to consuming modules.

* refactor: create temp folders for src and dist files

We copy the src users source folder to a temp dir. We do that so we can
compare the files in a time when the src folder might be overridden by
the parsed files

* refactor(parse-all): report error per parser

When a parser errors, report what happend to the user. Continue with all
parsers before exiting. Parsers are started in paralell, we need to wait
for all of them to complete to clean up the temp folder.

* refactor(index): return the promise chain

By returning the promise chain consuming modules can act on the
resolving or rejecting of the promise chain. As a result, the
bin/fastatic module can set the exitCode as discussed in #75