Filesize Reporter module (#94)
* refactor(parse-all): report error per parser

When a parser errors, report what happend to the user. Continue with all
parsers before exiting. Parsers are started in paralell, we need to wait
for all of them to complete to clean up the temp folder.

* refactor(index): return the promise chain

By returning the promise chain consuming modules can act on the
resolving or rejecting of the promise chain. As a result, the
bin/fastatic module can set the exitCode as discussed in #75

* refactor(index): split execution from cleanup flow

To be able to return the result to consuming modules, the result of the
parsing should be returned before the cleanup is executed

* feat(results): add results module

The results module returns an object with stats. The fastatic module
returns this object to consuming modules

* refactor(results): split functions in smaller pieces

* refactor(results): remove saved bytes from result

* refactor(results): rename to compareFileSize

* refactor(compare-file-size): accept two dirs instead of config object

* refactor(compare-file-size): return sizes object per folder

* refactor(compare-file-size): put calculating file sizes in own module

* refactor(index): return an object with the fileSizes as a property

Now we can output more data to consuming modules

* fix(index): replace console log with the logger module

* refactor: create temp folders for src and dist files

We copy the src users source folder to a temp dir. We do that so we can
compare the files in a time when the src folder might be overridden by
the parsed files

* refactor(index): remove unneeded Promise.all

* refactor(compare-file-sze): use path module to append dir to files

* refactor(get-filesizes-per-type): simplify variable names

* refactor(get-filesizes-per-type): simplify size addition

* refactor(parse-all): remove unused dependencies

* fix(get-filesizes-per-type): fix variable naming collisions

* chore: remove unneeded white space

* refactor(compare-file-size): keep variable names consistant

* style: add or remove extra tab

* feat(reporter): add reporter skeleton module

* feat(logger): add updateLevel method to logger

The loglevel should be updateable during runtime

* chore(package): add glob as a dependency

* refactor(stats-reporter): generate table based on stats data

* refactor(index): remove old stats module

* refactor(reporter): readd line indicating where the files are stored

* chore: remove unused dependencies

* refactor(stats-reporter): rename to filesizeReporter

* chore(reporter): rename folder to reporters

* refactor(fastatic): put filesizereporter module directly in fastatic

* refactor(filesize-reporter): pass loglevel to reporter

* refactor(filesize-reporter): calculate the savings in getSizesPerType

The dedicated calculateSavingsPerType function is a bit overkill

* refactor(filesize-reporter): rename getSizesPerType function

* refactor(filesize-reporter): use consistant naming

* refactor(filesize-reporter): let createTable walk over data

Instead of a dedicated createTableRows method, combine the logic in the
createTable method

* refactor(filesize-reporter): rename method for consistancy

* refactor(filesize-reporter): make functions smaller

* refactor(filesize-reporter): rename argument

* refactor(filesize-reporter): rename variables, put tableConfig ontop

* refactor(fastatic): keep var names consistant